Rip out other protocol support from gio
So it only supports file:// and local file access.
