+++ Packaging/debian/cp-src/libfontforge1.install
@@ -1,5 +1,4 @@
 usr/lib/libfontforge.so.*
 usr/lib/libfontforgeexe.so.*
-usr/lib/libgioftp.so.*
 usr/lib/libgunicode.so.*
 usr/lib/libgutils.so.*
