+++ doc/html/uninstall.html
@@ -20,7 +20,7 @@
     <PRE>$ su
 password:
 $ rm -rf /usr/local/bin/{fontforge,sfddiff,fontimage,fontlint} \
-/usr/local/lib/{libgdraw,libgunicode,libuninameslist,libspiro,libfontforge,libgutils,libgioftp}* \
+/usr/local/lib/{libgdraw,libgunicode,libuninameslist,libspiro,libfontforge,libgutils}* \
 /usr/local/lib/pkgconfig/fontforge.pc \
 /usr/local/man/man1/{fontforge,sfddiff,fontimage,fontlint}.1 \
 /usr/local/share/doc/fontforge \
@@ -33,7 +33,7 @@ $ rm -rf /usr/local/bin/{fontforge,sfddiff,fontimage,fontlint} \
   or if you are on a mac:
   <BLOCKQUOTE>
     <PRE>$ sudo rm -rf /usr/local/bin/{fontforge,sfddiff,fontimage,fontlint} \
-/usr/local/lib/{libgdraw,libgunicode,libuninameslist,libspiro,libfontforge,libgutils,libgioftp}* \
+/usr/local/lib/{libgdraw,libgunicode,libuninameslist,libspiro,libfontforge,libgutils}* \
 /usr/local/lib/pkgconfig/fontforge.pc \
 /usr/local/man/man1/{fontforge,sfddiff,fontimage,fontlint}.1 \
 /usr/local/share/doc/fontforge \
@@ -48,7 +48,7 @@ password:
   or if you are on cygwin:
   <BLOCKQUOTE>
     <PRE>$ rm -rf /usr/local/bin/{fontforge,sfddiff,fontimage,fontlint} \
-/usr/local/lib/{libgdraw,libgunicode,libuninameslist,libspiro,libfontforge,libgutils,libgioftp}* \
+/usr/local/lib/{libgdraw,libgunicode,libuninameslist,libspiro,libfontforge,libgutils}* \
 /usr/local/lib/pkgconfig/fontforge.pc \
 /usr/local/man/man1/{fontforge,sfddiff,fontimage,fontlint}.1 \
 /usr/local/share/doc/fontforge \
