+++ gutils/Makefile.am
@@ -35,8 +35,8 @@ libgutils_la_SOURCES = dlist.c fsys.c g_giomime.c gcol.c	\
 	gimagereadras.c gimagereadrgb.c gimagereadtiff.c		\
 	gimagereadxbm.c gimagereadxpm.c gimagewritebmp.c		\
 	gimagewritegimage.c gimagewritejpeg.c gimagewritepng.c		\
-	gimagewritexbm.c gimagewritexpm.c gio.c giofile.c giohosts.c	\
-	giothread.c giotrans.c giofuncP.h gioP.h gnetwork.c gutils.c	\
+	gimagewritexbm.c gimagewritexpm.c gio.c giofile.c		\
+	giotrans.c giofuncP.h  gnetwork.c gutils.c			\
 	gwwintl.c prefs.c prefs.h unicodelibinfo.c unicodelibinfo.h
 
 libgutils_la_CFLAGS = $(WARN_CFLAGS)
