+++ inc/gio.h
@@ -40,8 +40,6 @@ typedef struct giocontrol {
     unichar_t *origpath;		/* what the user asked for (before any redirects), NULL if path doesn't change */
     unichar_t *topath;			/* for renames and copies */
     void *userdata;
-    struct gio_connectiondata *connectiondata;
-    struct gio_threaddata *threaddata;
     void *iodata;
     void (*receivedata)(struct giocontrol *);
     void (*receiveintermediate)(struct giocontrol *);
@@ -93,14 +91,7 @@ extern void GIOclose(GIOControl *gc);
 extern GIOControl *GIOCreate(unichar_t *path,void *userdata,
 	void (*receivedata)(struct giocontrol *),
 	void (*receiveerror)(struct giocontrol *));
-extern void GIOSetDefAuthorizer(int32 (*getauth)(struct giocontrol *));
-extern void GIOSetUserAgent(unichar_t *agent);
 
-extern char *GIO_PasswordCache(char *proto,char *host,char *username,char *password);
-extern char *_GIO_decomposeURL(const unichar_t *url,char **host, int *port, char **username,
-	char **password);
-
-extern void GIO_SetThreadCallback(void (*callback)(void *,void *,void *));
 
 extern char *GIOguessMimeType(const char *path);
 extern char *GIOGetMimeType(const char *path);
